+++ pwchem/viewers/__init__.py
@@ -3,7 +3,5 @@ from .viewer_protocol_ZL_predict import ProtChemZLPredictViewer
 from .viewer_protocol_raptorX import ProtChemRaptorXViewer
 from .viewers_data import *
 from .viewer_pockets import *
-from .viewer_protocol_consensus_pockets import ViewerConsensusPockets
-from .viewer_smallMols import SmallMoleculesViewer
-from .viewer_docking import ProtConsensusDockingViewer
+from .viewer_smallMols import SmallMoleculesViewer, ProtConsensusDockingViewer
 from .viewers_sequences import SequenceAliViewer
